References/Microcontroller/control-lab/Python-Portable/requirements.txt
PyOpenGL
Pygame
pySerial
